+++ src/ClassicCalculator/CalculatorState/OperandInputNotInProgressState.cs
@@ -2,9 +2,9 @@
 {
     public class OperandInputNotInProgressState(
         ICalculator calculator,
-        double? firstOperand,
+        decimal? firstOperand,
         OperationType? currentOperation,
-        double? secondOperand,
+        decimal? secondOperand,
         string displayValue)
         : ValidStateBase(
             calculator,
